+++ src/cache.zig
@@ -12,7 +12,6 @@ const table_normalize_q =
 const fetch_news_q =
     \\WITH query (id, feed, title, pub_date, read_no)
     \\AS (
-    \\
     \\    SELECT items.id as item_id,
     \\        feed.title as feed_title,
     \\        items.title as item_title,
@@ -22,12 +21,56 @@ const fetch_news_q =
     \\    JOIN rss_feed as feed on feed.rssurl = items.feedurl
     \\    WHERE datetime(items.pubDate, 'unixepoch') >= datetime('now', ?)
     \\    AND items.unread=1
-    \\    
     \\)
     \\UPDATE rss_item SET read_no=(
     \\    SELECT MIN(read_no) from query WHERE read_no > (
     \\      SELECT MIN(query.read_no) from query)
     \\    ) + 1
+    \\WHERE rss_item.id IN (
+    \\    SELECT id FROM query
+    \\    WHERE query.read_no=(SELECT MIN(query.read_no) FROM query)
+    \\    ORDER BY query.pub_date DESC
+    \\    LIMIT 1
+    \\)
+    \\ RETURNING rss_item.id, rss_item.title, rss_item.url, rss_item.read_no;
+;
+
+const fetch_news_q_random =
+    \\WITH query (id, feed, title, pub_date, read_no)
+    \\AS (
+    \\    SELECT items.id as item_id,
+    \\        feed.title as feed_title,
+    \\        items.title as item_title,
+    \\        items.pubDate as pub_date,
+    \\        items.read_no as read_no
+    \\        FROM rss_item as items
+    \\    JOIN rss_feed as feed on feed.rssurl = items.feedurl
+    \\    WHERE datetime(items.pubDate, 'unixepoch') >= datetime('now', ?)
+    \\    AND items.unread=1
+    \\)
+    \\UPDATE rss_item SET read_no=read_no+1
+    \\WHERE rss_item.id IN (
+    \\    SELECT id FROM query
+    \\    ORDER BY RANDOM()
+    \\    LIMIT 1
+    \\)
+    \\ RETURNING rss_item.id, rss_item.title, rss_item.url, rss_item.read_no;
+;
+
+const fetch_news_q_latest =
+    \\WITH query (id, feed, title, pub_date, read_no)
+    \\AS (
+    \\    SELECT items.id as item_id,
+    \\        feed.title as feed_title,
+    \\        items.title as item_title,
+    \\        items.pubDate as pub_date,
+    \\        items.read_no as read_no
+    \\        FROM rss_item as items
+    \\    JOIN rss_feed as feed on feed.rssurl = items.feedurl
+    \\    WHERE datetime(items.pubDate, 'unixepoch') >= datetime('now', ?)
+    \\    AND items.unread=1
+    \\)
+    \\UPDATE rss_item SET read_no=read_no+1
     \\WHERE rss_item.id IN (
     \\    SELECT id FROM query
     \\    WHERE query.read_no=(SELECT MIN(query.read_no) FROM query)
@@ -88,25 +131,68 @@ pub const ArticleResult = Tuple(&.{
     []const u8,
 });
 
+fn getFetchArticleStmt(db: *sqlite.Db) !sqlite.DynamicStatement {
+    std.log.debug("Retrieving article using normal strategy", .{});
+    var diags = sqlite.Diagnostics{};
+    const stmt = db.prepareDynamicWithDiags(
+        fetch_news_q,
+        .{ .diags = &diags },
+    ) catch |err| {
+        std.log.err("Stmt error, err {}. diag: {s}", .{ err, diags });
+        return err;
+    };
+    return stmt;
+}
+
+fn getFetchArticleStmtRandom(db: *sqlite.Db) !sqlite.DynamicStatement {
+    std.log.debug("Retrieving article using random strategy", .{});
+    var diags = sqlite.Diagnostics{};
+    const stmt = db.prepareDynamicWithDiags(
+        fetch_news_q_random,
+        .{ .diags = &diags },
+    ) catch |err| {
+        std.log.err("Stmt error, err {}. diag: {s}", .{ err, diags });
+        return err;
+    };
+    return stmt;
+}
+
+fn getFetchArticleStmtLatest(db: *sqlite.Db) !sqlite.DynamicStatement {
+    std.log.debug("Retrieving article using latest strategy", .{});
+    var diags = sqlite.Diagnostics{};
+    const stmt = db.prepareDynamicWithDiags(
+        fetch_news_q_random,
+        .{ .diags = &diags },
+    ) catch |err| {
+        std.log.err("Stmt error, err {}. diag: {s}", .{ err, diags });
+        return err;
+    };
+    return stmt;
+}
+
 pub fn fetchArticle(
     db: *sqlite.Db,
     allocator: std.mem.Allocator,
     max_age: u16,
+    random: bool,
+    latest: bool,
 ) !?ArticleResult {
-    var diags = sqlite.Diagnostics{};
-    var stmt = db.prepareWithDiags(
-        fetch_news_q,
-        .{ .diags = &diags },
-    ) catch |err| {
-        std.log.err("unable to prepare statement, got error {}. diagnostics: {s}", .{ err, diags });
-        return err;
-    };
+    var stmt_maybe: ?sqlite.DynamicStatement = null;
+    if (random) {
+        stmt_maybe = try getFetchArticleStmtRandom(db);
+    } else if (latest) {
+        stmt_maybe = try getFetchArticleStmtLatest(db);
+    } else {
+        stmt_maybe = try getFetchArticleStmt(db);
+    }
+    var stmt = stmt_maybe.?;
     defer stmt.deinit();
     const t = try std.fmt.allocPrint(
         allocator,
         "-{d} hours",
         .{max_age},
     );
+    defer allocator.free(t);
     const row = try stmt.oneAlloc(
         struct {
             id: usize,
@@ -119,9 +205,17 @@ pub fn fetchArticle(
         .{ .t = t },
     );
     if (row) |r| {
+        std.log.debug("Row: {d};{s};{s};{d}", .{
+            r.id,
+            r.title,
+            r.url,
+            r.read_no,
+        });
         const title = try allocator.dupe(u8, r.title);
         const url = try allocator.dupe(u8, r.url);
         return .{ title, url };
+    } else {
+        std.log.debug("Empty row", .{});
     }
     return null;
 }