+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add --random and --latest options for headline retrieval
 ## [0.2.1] 2025-12-09
 ### Changed
 - Refactor setting read_no for newest articles